Introduce environment variable support
Here we introduce a small change to allow for environment variables.
Many hosting platforms will indicate a random port for an application to
leverage and with the introduction of the `startup.sh` script I want to
make that a bit more robust.

To this end there is now a `env.example` file to help as well as new
documentation in the README and a new Dockerfile directive to make sure
these variables are available to Docker.